deps/torch_robotics/requirements.txt
pyyaml
urdf-parser-py
matplotlib
trimesh
einops
